test: address review feedback on introspection coverage
AM-7456
