+++ gravitee-am-ui/build.json
@@ -1,3 +1,3 @@
 {
-  "version": "4.13.0-alpha.3"
+  "version": "4.13.0-alpha.4-SNAPSHOT"
 }
