+++ gravitee-am-ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "gravitee-am-webui",
-  "version": "4.13.0-alpha.3",
+  "version": "4.13.0-alpha.4-SNAPSHOT",
   "license": "Apache-2.0",
   "scripts": {
     "ng": "ng",
